embedding_consumer/requirements.txt
pgvecto_rs[sdk]==0.2.2
langchain-community==0.3.14
python-dotenv==1.0.1
coloredlogs==15.0.1
psycopg[binary]==3.2.3
langchain_ollama==0.2.2
langchain-elasticsearch==0.3.0
aiokafka==0.12.0
fastapi==0.115.7
uvicorn==0.34.0